refactor: move Tab type to types folder
