+++ src/types/tab.ts
@@ -1,0 +1,7 @@
+import { FieldValue } from "firebase/firestore";
+export interface Tab extends chrome.tabs.Tab {
+  lastAccessed: FieldValue;
+  spaceId?: string;
+  tabId: number | undefined;
+  isPinned: boolean;
+}
